refactor : member domain 병합 완료
