+++ src/main/java/com/rualone/app/domain/auth/application/OAuthLoginService.java
@@ -3,8 +3,8 @@ package com.rualone.app.domain.auth.application;
 import com.rualone.app.domain.auth.dto.infoResponse.OAuthInfoResponse;
 import com.rualone.app.domain.auth.params.OAuthLoginParams;
 import com.rualone.app.domain.auth.tokens.AuthTokensGenerator;
-import com.rualone.app.domain.memberOrigin.entity.People;
-import com.rualone.app.domain.memberOrigin.dao.PeopleRepository;
+import com.rualone.app.domain.member.entity.Member;
+import com.rualone.app.domain.member.dao.MemberRepository;
 import com.rualone.app.domain.auth.tokens.AuthTokens;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 public class OAuthLoginService {
-    private final PeopleRepository memberRepository;
+    private final MemberRepository memberRepository;
     private final AuthTokensGenerator authTokensGenerator;
     private final RequestOAuthInfoService requestOAuthInfoService;
 
@@ -23,24 +23,24 @@ public class OAuthLoginService {
     public AuthTokens login(OAuthLoginParams params) {
         OAuthInfoResponse oAuthInfoResponse = requestOAuthInfoService.request(params);
 
-        People people = findOrCreateMember(oAuthInfoResponse);
+        Member member = findOrCreateMember(oAuthInfoResponse);
 
-        AuthTokens authTokens = authTokensGenerator.generate(people.getId());
-        people.updateRefreshToken(authTokens.getRefreshToken());
+        AuthTokens authTokens = authTokensGenerator.generate(member.getId());
+        member.updateRefreshToken(authTokens.getRefreshToken());
 
         return authTokens;
     }
-    private People findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
+    private Member findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
         return memberRepository.findByEmail(oAuthInfoResponse.getEmail())
                 .orElseGet(() -> newMember(oAuthInfoResponse));
     }
 
-    private People newMember(OAuthInfoResponse oAuthInfoResponse) {
-        People people = People.builder()
+    private Member newMember(OAuthInfoResponse oAuthInfoResponse) {
+        Member member = Member.builder()
                 .email(oAuthInfoResponse.getEmail())
-                .nickname(oAuthInfoResponse.getNickname())
+                .nickName(oAuthInfoResponse.getNickname())
                 .oAuthProvider(oAuthInfoResponse.getOAuthProvider())
                 .build();
-        return memberRepository.save(people);
+        return memberRepository.save(member);
     }
 }
