+++ src/main/java/com/rualone/app/domain/board/api/CommentApi.java
@@ -5,8 +5,8 @@ import com.rualone.app.domain.board.dto.response.CommentResponse;
 import com.rualone.app.domain.board.entity.Comment;
 import com.rualone.app.domain.board.application.CommentService;
 import com.rualone.app.domain.board.dto.request.CommentCreateRequest;
-import com.rualone.app.domain.memberOrigin.application.MemberService;
-import com.rualone.app.domain.memberOrigin.entity.Member;
+import com.rualone.app.domain.member.application.MemberService;
+import com.rualone.app.domain.member.entity.Member;
 import com.rualone.app.global.api.ApiResult;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
