+++ src/main/java/com/rualone/app/domain/board/api/PostApi.java
@@ -7,8 +7,8 @@ import com.rualone.app.domain.board.dto.request.PostCreateRequest;
 import com.rualone.app.domain.board.dto.request.PostUpdateRequest;
 import com.rualone.app.domain.board.dto.response.PostDetailResponse;
 import com.rualone.app.domain.board.dto.response.PostResponse;
-import com.rualone.app.domain.memberOrigin.application.MemberService;
-import com.rualone.app.domain.memberOrigin.entity.Member;
+import com.rualone.app.domain.member.application.MemberService;
+import com.rualone.app.domain.member.entity.Member;
 import com.rualone.app.global.api.ApiResult;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
