+++ src/main/java/com/rualone/app/domain/board/application/impl/CommentServiceImpl.java
@@ -8,7 +8,7 @@ import com.rualone.app.domain.board.dto.request.CommentCreateRequest;
 import com.rualone.app.domain.board.entity.Post;
 import com.rualone.app.domain.board.validator.CommentValidator;
 import com.rualone.app.domain.board.validator.PostValidator;
-import com.rualone.app.domain.memberOrigin.entity.Member;
+import com.rualone.app.domain.member.entity.Member;
 import com.rualone.app.global.error.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
