+++ src/main/java/com/rualone/app/domain/board/application/impl/PostLikeServiceImpl.java
@@ -5,8 +5,8 @@ import com.rualone.app.domain.board.dao.PostLikeRepository;
 import com.rualone.app.domain.board.entity.Post;
 import com.rualone.app.domain.board.entity.PostLike;
 import com.rualone.app.domain.board.validator.PostValidator;
-import com.rualone.app.domain.memberOrigin.entity.Member;
-import com.rualone.app.domain.memberOrigin.validator.MemberValidator;
+import com.rualone.app.domain.member.entity.Member;
+import com.rualone.app.domain.member.validator.MemberValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
