+++ src/main/java/com/rualone/app/domain/member/dao/MemberRepository.java
@@ -1,0 +1,13 @@
+package com.rualone.app.domain.member.dao;
+
+import com.rualone.app.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+
+    Optional<Member> findById(Long id);
+}
+
