--- src/main/java/com/rualone/app/domain/memberOrigin/dto/request/MemberLoginRequest.java
@@ -1,22 +1,0 @@
-package com.rualone.app.domain.memberOrigin.dto.request;
-
-import com.rualone.app.domain.memberOrigin.entity.Member;
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class MemberLoginRequest {
-    private String loginId;
-    private String password;
-
-    public Member ToEntity() {
-        return Member.builder()
-                .loginId(loginId)
-                .password(password)
-                .build();
-    }
-}
